refactor traverse content component
